tests/requirements.txt
hypothesis>=6.100.0
pyyaml>=6.0
pytest>=8.0
